Reviewed by Brady Eidson.
        Ninth round of fixes for implicit 64-32 bit conversion errors.
        <rdar://problem/5292262>

        - Fix function signature to use unsigned long instead of unsigned long long.

        * loader/ProgressTracker.cpp:
        (WebCore::ProgressTracker::completeProgress):
        * loader/ProgressTracker.h:



Canonical link: https://commits.webkit.org/18707@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@23953 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/loader/ProgressTracker.cpp
@@ -224,7 +224,7 @@ void ProgressTracker::incrementProgress(unsigned long identifier, const char*, i
     m_originatingProgressFrame->loader()->client()->didChangeEstimatedProgress();
 }
 
-void ProgressTracker::completeProgress(unsigned long long identifier)
+void ProgressTracker::completeProgress(unsigned long identifier)
 {
     ProgressItem* item = m_progressItems.get(identifier);
